backend/requirements.txt
fastapi==0.122.0
uvicorn==0.38.0
python-dotenv==1.2.1
pydantic==2.12.5
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
langchain==0.3.26
langchain-community==0.3.26
langchain-google-genai
google-generativeai==0.8.5
python-multipart==0.0.20

